// src/cars/interface/car.interface.ts
export interface Car {
  acceleration: string;
  brand: string;
  calendar: Date[];
  color: string;
  dailyPrice: string;
  description: string;
  drivetrain: string;
  engine: string;
  equipment: string[];
  gas: string;
  img: string[];
  max: string;
  mileage: string;
  name: string;
  power: string;
  priceList: Map<string, number>;
  seats: string;
  torque: string;
  transmission: string;
  year: string;
}